+++ src/validate/errors.js
@@ -1,0 +1,16 @@
+class ValidationError extends Error {
+  constructor(code, message) {
+    super(`Validation error [${code}]: ${message}`);
+    this.name = "ValidationError";
+    this.code = code;
+  }
+}
+
+function failValidation(code, message) {
+  throw new ValidationError(code, message);
+}
+
+module.exports = {
+  ValidationError,
+  failValidation,
+};